Rework of location car possiblities.
Consider the perspective of the location manager at the end of the
day: there are so many cars on the lot, what are all of the potential
ways in which this happend? In some cases, cars got moved from the
other location -- permutations at the other location should take this
into account.