Car counting logic closer to problem interpretation.
The way I currently read the problem, Jack wants to know how many cars
to move after hours. Thus, the number of cars in the current state
means two things: whether the requested number of cars to move is
valid, and a basis for how business that day went.

This produces figures that are closer to what the book has, but does
so in two steps rather than five. The previous version of this file
had results that didn't seem correct, but did require the correct
number of steps to terminate.