Correct model for probability and transitions.
When given two states at a particular location, the old way of
thinking was to consider the difference as a function of the number of
potentially returned and/or rented. The correct way to think about
this is the number potentially requested and returned; where the
number rented is a function of returned and the state.

This change was inspired by reading various implementations online,
more notably: https://github.com/naharabhishek/Reinforcement-Learning

Determining rewards and transitions probabilities are now the
responsibility of the Location.