Break ties.
It is possible that there are several policies that are optimal, we
should thus keep them all and break the ties in some way. Based on
Figure 4.6 (lower), it appears that the book breaks ties by taking the
smallest action of those available -- that is replicated here.

The final policy graph is still not what is in the book, however. This
is likely due to floating point errors.