Free servers to index mapping responsibility of policy
The server class should only worry about the number of free servers,
not how they are indexed. Adding one to the number specified, and
having the System keep track of what was returned, is
awkward. Ultimately, the servers are just identifiers; what the policy
keeps track of what those identifiers mean for free accountability.