Initial interpretation of n-Step TD
Example 7.1 is an extension of the random walk from the previous
chapter. As such, this module is an updated version of that, with
(intended) support for n-Step TD instead of 1-step and Monte
Carlo.

Unlike previous chapters in which the algorithm is laid out, n-Step is
presented in paragraph form; mainly over pages 165 and 166. This is an
attempt to capture that.